Fix tests to work with new arch
